Added angelr_main.py
angelr_main.py consists of all the classes Review, Reply, Restaurant, Owner, and Customer along with most of the methods and workflows. The main menu loop is also included at the bottom where the user can navigate. As of now I forgot to give the owner to add a restaurant in the owner management menu but just implemented that when a new account is created as an owner they are asked if they want to initially add a restaurant but I need to add the options to add/delete/modify the restaurant once they login and are in the owner management menu